% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsetcor.R
\name{cor_pvalue}
\alias{cor_pvalue}
\title{Calculate p-value from correlation rho}
\usage{
cor_pvalue(x, n)
}
\arguments{
\item{x}{correlation value}

\item{n}{degrees of freendom}
}
\value{
p valu
}
\description{
Calculate p-value from correlation rho
}
