% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFET.R
\name{corpora.fisher.pval}
\alias{corpora.fisher.pval}
\title{Superfast version of Fisher Exact Test from 'corpora' R package.}
\usage{
corpora.fisher.pval(
  k1,
  n1,
  k2,
  n2,
  alternative = c("two.sided", "less", "greater"),
  log.p = FALSE
)
}
\description{
Superfast version of Fisher Exact Test from 'corpora' R package.
}
