% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFET.R
\name{corporaFET}
\alias{corporaFET}
\title{Wrapper around fast version of Fisher Exact Test from 'corpora' R
package.}
\usage{
corporaFET(a, b, c, d)
}
\description{
Wrapper around fast version of Fisher Exact Test from 'corpora' R
package.
}
