% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultragsea.R
\name{fc_zmat}
\alias{fc_zmat}
\title{Compute z-score matrix for gene sets}
\usage{
fc_zmat(fc, G, alpha = 0.5)
}
\arguments{
\item{fc}{Numeric vector of (log)foldchanges, with genes as names}

\item{G}{Numeric matrix of gene sets, with genes as row names}

\item{alpha}{Numeric value between 0 and 1 for Bayesian shrinkage}
}
\value{
Z-score matrix
}
\description{
Compute z-score matrix for gene sets
}
\keyword{internal}
