% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultraGSEA.R
\name{fc_ztest}
\alias{fc_ztest}
\title{Fast one sample z-test for matrix object F (e.g. foldchanges) and
grouping matrix G (e.g. gene sets).}
\usage{
fc_ztest(fc, G, zmat = FALSE, alpha = 0.5)
}
\description{
Fast one sample z-test for matrix object F (e.g. foldchanges) and
grouping matrix G (e.g. gene sets).
}
