% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultragsea.R
\name{fgsea}
\alias{fgsea}
\title{Fast replacement of fgsea with p/q values computed with ultragsea.cor}
\usage{
fgsea(pathways, stats, minSize = 1, maxSize = length(stats) - 1, nperm = 20)
}
\description{
Fast replacement of fgsea with p/q values computed with ultragsea.cor
}
