% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsetcor.R
\name{gset.cor}
\alias{gset.cor}
\title{Calculate gene set correlation}
\usage{
gset.cor(FC, gset, compute.p = FALSE, use.rank = FALSE)
}
\arguments{
\item{FC}{Numeric vector or matrix of (log)foldchanges, with genes as row names}

\item{gset}{Numeric matrix of gene sets, with genes as row/column names}

\item{compute.p}{Logical indicating whether to compute p-values}

\item{use.rank}{Logical indicating whether to rank transform FC before correlation}
}
\value{
Named list with components:
\itemize{
\item rho - Matrix of correlation coefficients between FC and gset
\item p.value - Matrix of p-values for correlation (if compute.p = TRUE)
\item q.value - Matrix of FDR adjusted p-values (if compute.p = TRUE)
}
}
\description{
Compute correlation between a foldchange vector/matrix and gene sets
}
\details{
This function calculates sparse rank correlation between FC and each
column of gset using \code{qlcMatrix::corSparse()}. It handles missing values in
FC by computing column-wise correlations.

P-values are computed from statistical distribution
}
\examples{
\dontrun{
librart(playbase)
ranks <- sample(1:10000, 1000, replace = TRUE)
names(ranks) <- replicate(1000, paste(sample(LETTERS, 4, replace = TRUE), collapse = ""))
genesets <- matrix(rnorm(1000 * 20), ncol = 20)
rownames(genesets) <- names(ranks)

gset.rankcor(ranks, genesets, compute.p = TRUE)
}
}
