% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFET.R
\name{gset.fastFET}
\alias{gset.fastFET}
\title{Calculate fast Fisher exact test.}
\usage{
gset.fastFET(
  genes,
  G,
  bg,
  report.genes = FALSE,
  method = c("phyper", "genesetr")[1]
)
}
\arguments{
\item{genes}{Vector of significant genes}

\item{G}{Sparse matrix containing gene sets}
}
\description{
Calculate fast Fisher exact test.
}
