% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultragsea.R
\name{gset.ztest}
\alias{gset.ztest}
\title{Fast one sample z-test for matrix object F (e.g. foldchanges) and
grouping matrix G (e.g. gene sets).}
\usage{
gset.ztest(F, G, alpha = 0.5, center = TRUE, pdist = "norm")
}
\arguments{
\item{G}{Numeric matrix of gene sets, with genes as row names}

\item{alpha}{Numeric value between 0 and 1 for Bayesian shrinkage}

\item{center}{Logical value to center population mean or assume zero}

\item{fc}{Numeric vector of (log)foldchanges, with genes as names}
}
\value{
List with components:
\itemize{
\item z_statistic - Vector of z-test statistics
\item p_value - Vector of p-values
}
}
\description{
Fast one sample z-test for matrix object F (e.g. foldchanges) and
grouping matrix G (e.g. gene sets).
}
