% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeGseaResult}
\alias{makeGseaResult}
\title{Utility to convert fgsea/ultragsea style results to 'gseaResult'
object compatible with clusterProfiler and enrichplot
functions. Using this function you can use all plotting functions
from enrichplot.}
\usage{
makeGseaResult(res, geneList, geneSets)
}
\description{
Utility to convert fgsea/ultragsea style results to 'gseaResult'
object compatible with clusterProfiler and enrichplot
functions. Using this function you can use all plotting functions
from enrichplot.
}
