% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultraGSEA.R
\name{ultraGSEA}
\alias{ultraGSEA}
\title{Ultra-fast GSEA using z-score or geneset correlation. Results of
these methods highly correlate with GSEA/fGSEA but are much faster.}
\usage{
ultraGSEA(
  fc,
  G,
  alpha = 0.5,
  minLE = 1,
  method = c("ztest", "ttest", "cor", "rankcor")[1],
  format = c("simple", "as.gsea")[1]
)
}
\description{
Ultra-fast GSEA using z-score or geneset correlation. Results of
these methods highly correlate with GSEA/fGSEA but are much faster.
}
