% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultragsea.R
\name{ultragsea}
\alias{ultragsea}
\title{Ultra-fast GSEA using z-score or geneset correlation. Results of
these methods highly correlate with GSEA/fGSEA but are much faster.}
\usage{
ultragsea(
  fc,
  G,
  alpha = 0.5,
  minLE = 1,
  corshrink = 3,
  minsize = 1L,
  maxsize = 9999L,
  center = TRUE,
  method = c("cor", "ztest", "ttest", "goat")[1],
  format = c("simple", "as.gsea", "as.gsea2")[1]
)
}
\description{
Ultra-fast GSEA using z-score or geneset correlation. Results of
these methods highly correlate with GSEA/fGSEA but are much faster.
}
